// src/BLL/Enums/Interval.enum.ts
export enum Interval {
    OneMinute = 0,
    ThreeMinutes = 1,
    FiveMinutes = 2,
    FifteenMinutes = 3,
    ThirtyMinutes = 4,
    OneHour = 5,
    TwoHour = 6,
    FourHour = 7,
    SixHour = 8,
    EightHour = 9,
    TwelveHour = 10,
    OneDay = 11,
    ThreeDay = 12,
    OneWeek = 13,
    OneMonth = 14
}